+++ django_fabfile/backup.py
@@ -20,12 +20,11 @@ from fabric.contrib.files import append
 from django_fabfile.instances import (attach_snapshot, create_temp_inst,
                                       get_avail_dev, get_vol_dev, mount_volume)
 from django_fabfile.utils import (
-    Config, StateNotChangedError, add_tags, config_temp_ssh,
+    StateNotChangedError, add_tags, config, config_temp_ssh,
     get_descr_attr, get_inst_by_id, get_region_conn, get_snap_device,
     get_snap_time, get_snap_vol, timestamp, wait_for, wait_for_sudo)
 
 
-config = Config()
 USERNAME = config.get('DEFAULT', 'USERNAME')
 env.update({'user': USERNAME, 'disable_known_hosts': True})